+++ conf/conf.go
@@ -2,19 +2,14 @@ package conf
 
 var (
 	// build information: this is set at compile time using LDFlags
-	buildVer    = "n/a"
-	buildDate   = "n/a"
-	buildCommit = "n/a"
+	buildVer  = "n/a"
+	buildDate = "n/a"
 )
 
 func BuildVersion() string {
 	return buildVer
 }
 
-func BuildData() string {
+func BuildDate() string {
 	return buildDate
-}
-
-func BuildCommit() string {
-	return buildCommit
 }